Add DeployInstanceConfigurationFromTemplates
